move NormalizingStreamV8 to its own class
