+++ liquibase-standard/src/main/java/liquibase/change/core/CreateViewChange.java
@@ -271,7 +271,7 @@ public class CreateViewChange extends AbstractChange {
                 }
             }
 
-            CheckSum checkSum = CheckSum.compute(new AbstractSQLChange.NormalizingStreamV8(";", false, false, stream), false, version);
+            CheckSum checkSum = CheckSum.compute(new NormalizingStreamV8(";", false, false, stream), false, version);
 
             return CheckSum.compute(super.generateCheckSum(version).toString() + ":" + checkSum, version);
         } finally {
